+++ script/compiled/combat/combat.js
@@ -203,11 +203,7 @@ function defeatedEnemies() {
     });
     text += `${Math.floor(combat.gold * player.allModifiers["goldGainP"])}<c>gold<c> ${game.getLocalizedString("gold")}\n`;
     text += `<c>silver<c>${Math.floor(combat.xp * player.allModifiers["expGainP"])}<c>lime<c> ${game.getLocalizedString("xp")}`;
-    if (currentStage.score > 0) {
-        const score = Math.floor(currentStage.score * scoreMultiplier());
-        text += `\n\n§<f>1.5rem<f><c>silver<c>${game.getLocalizedString("score_received").replace("{score}", score.toString())}`;
-    }
-    return textSyntax(text);
+    return text;
 }
 function lootEnemies(enemies) {
     let total = [];
@@ -289,7 +285,21 @@ class Combat {
             combatSummaryTitle.innerText = game.getLocalizedString("combat_victory");
             combatSummaryTitle.classList.value = "header victory";
             if (canGetLoot) {
-                combatSummaryText.append(defeatedEnemies());
+                let text = defeatedEnemies();
+                let score = 0;
+                if (dungeonController.currentRoom?.id) {
+                    if (!player.hasCompletedRoom(dungeonController.currentRoom.id)) {
+                        score = dungeonController.currentRoom.score || 0;
+                    }
+                }
+                else if (!player.hasCompletedStage(currentStage.id)) {
+                    score = currentStage.score || 0;
+                }
+                if (score > 0) {
+                    score = Math.floor(score * scoreMultiplier());
+                    text += `\n\n§<f>1.5rem<f><c>silver<c>${game.getLocalizedString("score_received").replace("{score}", score.toString())}`;
+                }
+                combatSummaryText.append(textSyntax(text));
             }
             combatSummaryButtons.innerHTML = `<button class="main-button" onclick="combat.finish_combat()">${game.getLocalizedString("continue")}</button>`;
         }
@@ -324,6 +334,7 @@ class Combat {
             : true;
         if (this.defeat) {
             player.xp -= this.xp;
+            stats.total_xp_lost += this.xp;
             if (challenge("hardcore")) {
                 this.playing = false;
                 mainMenuElement.classList.remove("no-display");